Implement beginning and ending stack and queue nodes to linked list to check mode.
